I'm nearly finished
The sites practically done.
The only issue is the sidebar it wont expand when the page overflows.
its a nice looking page. I think the quality that saved it is the combination of the sidebar font and color.
I need to take UI/UX courses though. Its a glaring flaw in all my projects.